Moved monitorStack to shared lib/plugins/aws/lib
