+++ lib/plugins/aws/deploy/index.js
@@ -2,6 +2,7 @@
 
 const BbPromise = require('bluebird');
 const validate = require('../lib/validate');
+const monitorStack = require('../lib/monitorStack');
 const createStack = require('./lib/createStack');
 const mergeCustomProviderResources = require('./lib/mergeCustomProviderResources');
 const generateArtifactDirectoryName = require('./lib/generateArtifactDirectoryName');
@@ -9,7 +10,6 @@ const setBucketName = require('./lib/setBucketName');
 const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
 const uploadArtifacts = require('./lib/uploadArtifacts');
 const updateStack = require('./lib/updateStack');
-const monitorStack = require('./lib/monitorStack');
 
 const SDK = require('../');
