+++ lib/plugins/aws/deploy/tests/all.js
@@ -7,5 +7,4 @@ require('./setBucketName');
 require('./cleanupS3Bucket');
 require('./uploadArtifacts');
 require('./updateStack');
-require('./monitorStack');
 require('./index');
