+++ lib/plugins/aws/lib/monitorStack.js
@@ -95,6 +95,8 @@ module.exports = {
               })
               .catch((e) => {
                 if (action === 'removal' && e.message.endsWith('does not exist')) {
+                  // empty console.log for a prettier output
+                  if (!this.options.verbose) this.serverless.cli.consoleLog('');
                   this.serverless.cli.log(`Stack ${action} finished...`);
                   resolve('DELETE_COMPLETE');
                 } else {
