+++ lib/plugins/aws/remove/index.js
@@ -2,9 +2,10 @@
 
 const BbPromise = require('bluebird');
 const validate = require('../lib/validate');
+const monitorStack = require('../lib/monitorStack');
 const emptyS3Bucket = require('./lib/bucket');
 const removeStack = require('./lib/stack');
-const monitorStack = require('../deploy/lib/monitorStack');
+
 const SDK = require('../');
 
 class AwsRemove {
