+++ lib/plugins/aws/tests/monitorStack.js
@@ -3,49 +3,52 @@
 const expect = require('chai').expect;
 const sinon = require('sinon');
 const BbPromise = require('bluebird');
-const AwsDeploy = require('../index');
-const Serverless = require('../../../../Serverless');
+const Serverless = require('../../../Serverless');
+const SDK = require('../index');
+const CLI = require('../../../classes/CLI');
+const monitorStack = require('../lib/monitorStack');
 
 describe('monitorStack', () => {
-  let serverless;
-  let awsDeploy;
+  const serverless = new Serverless();
+  const awsPlugin = {};
 
   beforeEach(() => {
-    serverless = new Serverless();
-    const options = {
+    awsPlugin.serverless = serverless;
+    awsPlugin.sdk = new SDK(serverless);
+    awsPlugin.serverless.cli = new CLI(serverless);
+    awsPlugin.options = {
       stage: 'dev',
       region: 'us-east-1',
     };
-    awsDeploy = new AwsDeploy(serverless, options);
 
-    awsDeploy.serverless.cli = new serverless.classes.CLI();
+    Object.assign(awsPlugin, monitorStack);
   });
 
   describe('#monitorStack()', () => {
     it('should skip monitoring if the --noDeploy option is specified', () => {
-      awsDeploy.options.noDeploy = true;
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      awsPlugin.options.noDeploy = true;
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
 
-      return awsDeploy.monitorStack('update', cfDataMock, 10).then(() => {
+      return awsPlugin.monitorStack('update', cfDataMock, 10).then(() => {
         expect(describeStackEventsStub.callCount).to.be.equal(0);
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should skip monitoring if the stack was already created', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
 
-      return awsDeploy.monitorStack('update', 'alreadyCreated', 10).then(() => {
+      return awsPlugin.monitorStack('update', 'alreadyCreated', 10).then(() => {
         expect(describeStackEventsStub.callCount).to.be.equal(0);
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should keep monitoring until CREATE_COMPLETE stack status', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -75,19 +78,19 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(0).returns(BbPromise.resolve(updateStartEvent));
       describeStackEventsStub.onCall(1).returns(BbPromise.resolve(updateFinishedEvent));
 
-      return awsDeploy.monitorStack('create', cfDataMock, 10).then((stackStatus) => {
+      return awsPlugin.monitorStack('create', cfDataMock, 10).then((stackStatus) => {
         expect(describeStackEventsStub.callCount).to.be.equal(2);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
         expect(stackStatus).to.be.equal('CREATE_COMPLETE');
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should keep monitoring until UPDATE_COMPLETE stack status', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -117,19 +120,19 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(0).returns(BbPromise.resolve(updateStartEvent));
       describeStackEventsStub.onCall(1).returns(BbPromise.resolve(updateFinishedEvent));
 
-      return awsDeploy.monitorStack('update', cfDataMock, 10).then((stackStatus) => {
+      return awsPlugin.monitorStack('update', cfDataMock, 10).then((stackStatus) => {
         expect(describeStackEventsStub.callCount).to.be.equal(2);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
         expect(stackStatus).to.be.equal('UPDATE_COMPLETE');
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should keep monitoring until DELETE_COMPLETE stack status', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -159,19 +162,19 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(0).returns(BbPromise.resolve(updateStartEvent));
       describeStackEventsStub.onCall(1).returns(BbPromise.resolve(updateFinishedEvent));
 
-      return awsDeploy.monitorStack('removal', cfDataMock, 10).then((stackStatus) => {
+      return awsPlugin.monitorStack('removal', cfDataMock, 10).then((stackStatus) => {
         expect(describeStackEventsStub.callCount).to.be.equal(2);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
         expect(stackStatus).to.be.equal('DELETE_COMPLETE');
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should keep monitoring until DELETE_COMPLETE or stack not found catch', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -193,20 +196,20 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(0).returns(BbPromise.resolve(updateStartEvent));
       describeStackEventsStub.onCall(1).returns(BbPromise.reject(stackNotFoundError));
 
-      return awsDeploy.monitorStack('removal', cfDataMock, 10).then((stackStatus) => {
+      return awsPlugin.monitorStack('removal', cfDataMock, 10).then((stackStatus) => {
         expect(describeStackEventsStub.callCount).to.be.equal(2);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
         expect(stackStatus).to.be.equal('DELETE_COMPLETE');
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should output all stack events information with the --verbose option', () => {
-      awsDeploy.options.verbose = true;
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      awsPlugin.options.verbose = true;
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -261,22 +264,22 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(2).returns(BbPromise.resolve(updateRollbackEvent));
       describeStackEventsStub.onCall(3).returns(BbPromise.resolve(updateRollbackFailedEvent));
 
-      return awsDeploy.monitorStack('update', cfDataMock, 10).catch((e) => {
+      return awsPlugin.monitorStack('update', cfDataMock, 10).catch((e) => {
         let errorMessage = 'An error occurred while provisioning your stack: ';
         errorMessage += 'mochaS3 - Bucket already exists.';
         expect(e.name).to.be.equal('ServerlessError');
         expect(e.message).to.be.equal(errorMessage);
         expect(describeStackEventsStub.callCount).to.be.equal(4);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should catch describeStackEvents error if stack was not in deleting state', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -286,19 +289,19 @@ describe('monitorStack', () => {
 
       describeStackEventsStub.onCall(0).returns(BbPromise.reject(failedDescribeStackEvents));
 
-      return awsDeploy.monitorStack('update', cfDataMock, 10).catch((e) => {
+      return awsPlugin.monitorStack('update', cfDataMock, 10).catch((e) => {
         expect(e.message).to.be.equal('Something went wrong.');
         expect(describeStackEventsStub.callCount).to.be.equal(1);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
 
     it('should throw an error if CloudFormation returned unusual stack status', () => {
-      const describeStackEventsStub = sinon.stub(awsDeploy.sdk, 'request');
+      const describeStackEventsStub = sinon.stub(awsPlugin.sdk, 'request');
       const cfDataMock = {
         StackId: 'new-service-dev',
       };
@@ -353,17 +356,17 @@ describe('monitorStack', () => {
       describeStackEventsStub.onCall(2).returns(BbPromise.resolve(updateRollbackEvent));
       describeStackEventsStub.onCall(3).returns(BbPromise.resolve(updateRollbackFailedEvent));
 
-      return awsDeploy.monitorStack('update', cfDataMock, 10).catch((e) => {
+      return awsPlugin.monitorStack('update', cfDataMock, 10).catch((e) => {
         let errorMessage = 'An error occurred while provisioning your stack: ';
         errorMessage += 'mochaS3 - Bucket already exists.';
         expect(e.name).to.be.equal('ServerlessError');
         expect(e.message).to.be.equal(errorMessage);
         expect(describeStackEventsStub.callCount).to.be.equal(4);
         expect(describeStackEventsStub.calledWith(
-          awsDeploy.options.stage,
-          awsDeploy.options.region
+          awsPlugin.options.stage,
+          awsPlugin.options.region
         ));
-        awsDeploy.sdk.request.restore();
+        awsPlugin.sdk.request.restore();
       });
     });
   });
